Extract kill_process to own module
This is mostly to make things a bit more separate since there’s a circular dependency, which is not great.
